whitelist/requirements.txt
requests==2.31.0
python-dotenv==1.0.1
mcrcon==0.7.0